synesis_cloud::tunnel::heartbeat

Struct HeartbeatService

pub struct HeartbeatService {
    config: HeartbeatConfig,
    sequence: Arc<AtomicU64>,
    connection: Arc<RwLock<Option<Connection>>>,
    shutdown: Sender<()>,
    prewarm_callback: Arc<RwLock<Option<PrewarmCallback>>>,
    last_prewarm: Arc<RwLock<Option<DateTime<Utc>>>>,
}

Heartbeat service

Sends periodic heartbeats to the cloud server with real device vitals

Fields
config: HeartbeatConfigsequence: Arc<AtomicU64>connection: Arc<RwLock<Option<Connection>>>shutdown: Sender<()>prewarm_callback: Arc<RwLock<Option<PrewarmCallback>>>last_prewarm: Arc<RwLock<Option<DateTime<Utc>>>>
Implementations

impl HeartbeatService

pub fn new(config: HeartbeatConfig) -> Self

Create a new heartbeat service

pub async fn set_connection(&self, conn: Connection)

Set the active connection

pub async fn clear_connection(&self)

Clear the connection (on disconnect)

pub async fn set_prewarm_callback(&self, callback: PrewarmCallback)

Set callback for pre-warm signals

pub fn spawn(&self) -> JoinHandle<()>

Spawn the heartbeat task

Returns a JoinHandle that can be used to wait for shutdown

async fn send_heartbeat( conn: &Connection, sequence: u64, vitals: DeviceVitals, ) -> CloudResult<HeartbeatAck>

Send a single heartbeat with real vitals

pub fn shutdown(&self)

Shutdown the heartbeat service